dom/payments/test/test_abortPayment.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1345367
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1345367</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script>
  SimpleTest.waitForExplicitFinish();

  var gUrl = SimpleTest.getTestFileURL('GeneralChromeScript.js');
  var gScript = SpecialPowers.loadChromeScript(gUrl);

  const defaultMethods = [{
    supportedMethods: "basic-card",
  }];
  const defaultDetails = {
    total: {
      label: "Total",
      amount: {
        currency: "USD",
        value: "1.00"
      }
    },
  };

  async function testBeforeShow() {
    const payRequest = new PaymentRequest(defaultMethods, defaultDetails);
    try {
      await payRequest.abort();
      ok(false, "Should throw 'InvalidStateError', but got resolved.");
    } catch (err) {
      console.log(err)
      is(err.name, "InvalidStateError", "Expected 'InvalidStateError', but got '" + err.name + "'");
    }
  }

  async function testAfterShow() {
    const handler = SpecialPowers.getDOMWindowUtils(window).setHandlingUserInput(true);
    const payRequest = new PaymentRequest(defaultMethods, defaultDetails);
    const acceptPromise = payRequest.show();
    try {
      const abortResult = await payRequest.abort();
      is(abortResult, undefined, "Should be resolved with undefined.");
    } catch (err) {
      ok(false, "Expected no error, but got '" + err.name + "'.");
    } finally {
      handler.destruct();
    }
  }

  function teardown() {
    return new Promise(resolve => {
      gScript.addMessageListener("teardown-complete", function teardownCompleteHandler() {
        gScript.removeMessageListener("teardown-complete", teardownCompleteHandler);
        gScript.destroy();
        resolve();
      });
      gScript.sendAsyncMessage("teardown");
    });
  }

  async function runTests() {
    try{
      await testBeforeShow();
      await testAfterShow();
    } catch(e) {
      ok(false, "Unexpected error: " + e.name);
    } finally {
      await teardown();
    }
  }

  window.addEventListener('load', async () => {
    await SpecialPowers.pushPrefEnv({
      'set': [
        ['dom.payments.request.enabled', true],
      ]
    });
    await runTests();
    SimpleTest.finish();
  });

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1345367">Mozilla Bug 1345367</a>
</pre>
</body>
</html>
